Fixed compile warning on GCC 2.95.3.
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@1557 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/MSXS1985.cc
@@ -27,7 +27,7 @@ byte MSXS1985::readIO(byte port, const EmuTime &time)
 	byte result;
 	switch (port) {
 		case 0x40:
-			result = ~ID;
+			result = (byte)~ID;
 			break;
 		case 0x42:
 			result = ram[address];
